clean up postgresql adapter.
